ci: fix lint (golangci-lint v2) + vuln-check (go1.25.10) to unblock Required Checks (#77)
* ci: fix lint (golangci-lint v2) and vuln-check (go1.25.10) so they pass

Required Checks was red on every Go PR — blocking the v1.0 stack — because
two CI-tooling jobs failed (not the code):

- Lint: golangci-lint v1 refuses to analyze a go1.25 module ("built with
  go1.24 < 1.25"). Bump golangci-lint-action to v7 (runs golangci-lint v2,
  built with a newer Go) and migrate .golangci.yml to v2. Run a pragmatic
  baseline (govet, bodyclose, unconvert) and DEFER the accumulated backlog
  (errcheck/unused/staticcheck/ineffassign — 50+ each — plus misspell, which
  mostly flags the pervasive, enum-bearing British "cancelled") and
  gofmt/goimports (the v2 binary's newer-Go formatter would rewrite ~250
  files). Tracked as a follow-up cleanup.
- Vuln-check: govulncheck flags 5 stdlib advisories
  (GO-2026-4976/4977/4980/4982/4986) fixed in go1.25.10. Pin
  `toolchain go1.25.10`.

Also fixes the handful of real findings the new baseline surfaced: drop
redundant conversions (json.RawMessage/int32/string) in remote_support_api,
health/enhanced, mfa/webauthn; mark the resilience HTTP wrapper's response
(returned to the caller) //nolint:bodyclose.

Validated locally with golangci-lint v2.12.2: `run ./...` = 0 issues; build OK.


* fix(deps): bump pgx/v5 to v5.5.4 to clear GO-2024-2606

With the toolchain pinned to go1.25.10 the 5 stdlib advisories cleared,
leaving one real, called dependency vulnerability: GO-2024-2606 in
github.com/jackc/pgx/v5 v5.5.3 (fixed in v5.5.4). Bump to v5.5.4 (patch).
govulncheck no longer reports it locally.


---------

+++ internal/health/enhanced.go
@@ -508,8 +508,8 @@ func gatherConnectionInfo(db *pgxpool.Pool, redisClient *redis.Client) *ConnInfo
 
 	if db != nil {
 		stats := db.Stat()
-		info.DBTotal = int32(stats.TotalConns())
-		info.DBIdle = int32(stats.IdleConns())
+		info.DBTotal = stats.TotalConns()
+		info.DBIdle = stats.IdleConns()
 		info.DBAcquired = stats.AcquireCount()
 	}
